ref(releases): Refactored Releases Serializers (#12535)
* refactored releases and added restriction that latest cannot be a ref or version name.

* moved release_head_commit to release.py because that's the only place it is used now.

* Fixed the missing fields.

* got org release details and org releases tests to work.

* ensured project release tests worked

* Added project release details and fixed the tests.

* removed ref check.
