+++ src/sentry/api/endpoints/project_releases.py
@@ -2,62 +2,20 @@ from __future__ import absolute_import
 
 from django.db import IntegrityError, transaction
 
-from rest_framework import serializers
 from rest_framework.response import Response
 
 from sentry.api.base import EnvironmentMixin
 from sentry.api.bases.project import ProjectEndpoint, ProjectReleasePermission
 from sentry.api.paginator import OffsetPaginator
-from sentry.api.fields.user import UserField
 from sentry.api.serializers import serialize
-from sentry.api.serializers.rest_framework import CommitSerializer, ListField
+from sentry.api.serializers.rest_framework import ReleaseWithVersionSerializer
 from sentry.models import (
     Activity,
-    CommitFileChange,
     Environment,
     Release,
 )
 from sentry.plugins.interfaces.releasehook import ReleaseHook
-from sentry.constants import VERSION_LENGTH
 from sentry.signals import release_created
-
-
-class CommitPatchSetSerializer(serializers.Serializer):
-    path = serializers.CharField(max_length=255)
-    type = serializers.CharField(max_length=1)
-
-    def validate_type(self, attrs, source):
-        value = attrs[source]
-        if not CommitFileChange.is_valid_type(value):
-            raise serializers.ValidationError('Commit patch_set type %s is not supported.' % value)
-        return attrs
-
-
-class CommitSerializerWithPatchSet(CommitSerializer):
-    patch_set = ListField(
-        child=CommitPatchSetSerializer(
-            required=False),
-        required=False,
-        allow_null=True)
-
-
-class ReleaseSerializer(serializers.Serializer):
-    version = serializers.CharField(max_length=VERSION_LENGTH, required=True)
-    ref = serializers.CharField(max_length=VERSION_LENGTH, required=False)
-    url = serializers.URLField(required=False)
-    owner = UserField(required=False)
-    dateReleased = serializers.DateTimeField(required=False)
-    commits = ListField(
-        child=CommitSerializerWithPatchSet(
-            required=False),
-        required=False,
-        allow_null=True)
-
-    def validate_version(self, attrs, source):
-        value = attrs[source]
-        if not Release.is_valid_version(value):
-            raise serializers.ValidationError('Invalid value for release')
-        return attrs
 
 
 class ProjectReleasesEndpoint(ProjectEndpoint, EnvironmentMixin):
@@ -146,7 +104,7 @@ class ProjectReleasesEndpoint(ProjectEndpoint, EnvironmentMixin):
                                       the current time is assumed.
         :auth: required
         """
-        serializer = ReleaseSerializer(data=request.DATA)
+        serializer = ReleaseWithVersionSerializer(data=request.DATA)
 
         if serializer.is_valid():
             result = serializer.object
